requirements.txt
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
cssselect2==0.9.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
freetype-py==2.5.1
html5lib==1.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pyHanko==0.34.0
pyhanko-certvalidator==0.30.0
PyJWT==2.11.0
pypdf==6.7.5
python-bidi==0.6.7
PyYAML==6.0.3
reportlab==4.4.10
requests==2.32.5
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tzdata==2025.3
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
xhtml2pdf==0.2.17
